content: add minimal JavaScript to expand examples
This is a followup to commit 97bb05a695b4a460d4cd3329cf4b9ebdbce11a21
that adds minimal code to expand and collapse documentation examples.
It incorporates the code review feedback from that CL.

The code may be simplified further in the future if it's viable to
reimplement collapsible examples using the <details> HTML element.

Change-Id: I6a722a62e6382776e8621b0200a35a923e86bd82
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/508137
Reviewed-by: Andrew Bonventre <andybons@google.com>