internal/dcensus,config: use distinct monitored resources for census export
Census export is updated to use new identities for monitored resources.
For metrics, the generic_task resource type is used to differentiate the
timeseries (gae_instance is not supported). For traces, the gae_app
resource type is used.

The config package is refactored to resolve all configuration at app
startup, and be more consistent in how configuration is accessed. Also,
query the metadata endpoint to resolve the GAE zone, which is needed for
the gae_app monitored resource.

Updates b/132629518

Change-Id: Ia8fd08c21f11271a1ea0841cdae15506157ae79b
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/509278
Reviewed-by: Julie Qiu <julieqiu@google.com>