internal/etl: implement special handling for builtin package
Package builtin in the standard library is a fake package that
provides documentation for Go's predeclared identifiers.
It requires special handling, such as displaying all identifiers
(not just exported ones), and not using type association.

This change implements that behavior and adds a test case for it.
Do validate documentation of packages in standard library, because
it's worth having test coverage and I'm not aware of any downsides.

Also update loadPackage documentation and specify that the
modulePath value for the standard library is "std".

Change-Id: I4bbc7af5c821be31fa975d3f1678c4a23b8bb1ff
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/509285
Reviewed-by: Julie Qiu <julieqiu@google.com>