internal/postgres: filter out empty license types for search
At the moment, if we don't detect a license type, then
the licenses.type column will be empty. These can lead to
an array containing the empty string to be populated in
mvw_search_documents (which will be fixed in a separate CL).

For now, filter out these empty license types when returning
search results.

Fixes b/137775128

Change-Id: Ie424cedbaf70efb2c8b480ebcf814d4843fc207c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/509860
Reviewed-by: Robert Findley <rfindley@google.com>