internal/frontend: reorganize details.go
details.go currently contains almost all of the logic for rendering the
package details view. In a future CL, it will also contain logic for
rendering the module details view.

It is reorganized by being split into multiple files in preparation for
handling that logic, which has a lot of shared functionality.

Updates b/138448402

Change-Id: Id031b14f58dce9e90b32ca658bfe920c30c42b74
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/515745
Reviewed-by: Robert Findley <rfindley@google.com>