discovery: implement initial module view
An intial version of the module view is implemented, which supports the
README and Packages tabs.

To do so, the following changes were made to the existing implementation
of the details page:

content,internal/frontend:
* fetchReadMeDetails and fetchModuleDetails now accept a *VersionInfo,
instead of a *VersionedPackage
* DetailsPage.TemplateName was added, to allow for specifying the
template that corresponds with a specific tab
* DetailsPage.Header was changed to an interface, so that either a
moduleHeader or packageHeader could be provided.
* templates specific to the package details page are now prefixed with "pkg_".

internal.postgres:
* GetVersionForPackage is renamed to GetPackagesForVersion, and the query
now accepts module path instead of package path as a param

Fixes b/138548892
Fixes b/138549609
Updates b/138448402

Change-Id: I58377a277536f39f9f453faef70cef4ae718d4b4
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/515746
Reviewed-by: Robert Findley <rfindley@google.com>