internal/proxy: require dot in stdlib prerelease versions
Go semver tags use prereleases like "-beta1". This is bad for sort order,
because, e.g. "-beta10" comes before "-beta9".

Make the proxy client require prereleases with dots, like "-beta.1".

Disallow "-beta1" and the like (return an error).

This also affects the zip files that the proxy client returns for the
stdlib. Their filenames will have the user prerelease ("-beta.1")
instead of the actual prerelease ("-beta1").  This seems to be the
intent of the createGoZipReader function.

Fixes b/138122236.

Change-Id: I51eec0eb77b0ebabd3a0df3a32a3b3f24cdc00fc
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/519424
Reviewed-by: Julie Qiu <julieqiu@google.com>