internal/postgres: fix errors with insert into search_documents
As of commit 6154707c8079a7e828779078dc21356942c41da1, a call to
InsertVersion inserted both to the versions and search_documents table.
We do not want internal packages to be inserted into search_documents,
but this error to db.UpsertSearchDocument was not handled properly in
InsertVersion. As a result, any module with an internal package now
fails to insert into the database. This error is now handled.

The logic for validating a version is also moved from saveVersion to
InsertVersion.

Change-Id: I1d411dfd632b97abec146c803eeb324e71b60f05
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/526390
Reviewed-by: Robert Findley <rfindley@google.com>