migrations: migration to add versions.series_path
When listing module versions, we want to include all versions
with the same series path.

The logic that extracts a series path from a module path is too
complex to encode in a query, and in any case that would duplicate the
logic, so we add a column to hold the series path.

Updates b/138547683.

Change-Id: Id4b0f673e125f743b19bfbfeb8eadd32d96bd719
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/529470
Reviewed-by: Robert Findley <rfindley@google.com>