internal/postgres: replace GetVersion with GetPackagesInVersion
We don't use the returned Version for anything other than its packages,
so changing the signature of this API both eliminates some joins, and
means that the Version type is only constructed by the fetch package.

Also update the implementation to use runQuery, and write to the Package
struct directly.

Fixes b/139976211
Updates b/130367504

Change-Id: I40884c27520017cd0fd79b285934e739d44c3df7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/535810
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>