internal: move licenses onto the Version struct
The slice of licenses used in the postgres.InsertVersion API is moved
directly on to the Version struct. The only reason licenses were *not*
contained on Version in the first place was to limit the Insert/Fetch
asymmetry in the postgres API. Since we no longer return any Versions
from postgres, this is not longer a problem. This CL simplifies the many
places in the code where we have to pass both Version and Licenses
together.

The meaning of the Version type is now 'the payload of data that we
extract from the module zip'. The type will be renamed to something more
descriptive in a later CL.

Updates b/139976128

Change-Id: Iba82d4363a93ffd3241bfe87cbf3d17f491de533
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/535812
Reviewed-by: Julie Qiu <julieqiu@google.com>