internal/proxy: add a wrapper for @latest endpoints
A method GetLatestInfo is added to the proxy client, to facilitate
lazily fetching the latest version of a package.

Additionally, a helper method executeRequest is added to handle making a
request to the proxy for a given URL. Its API is similar to runQuery in
the postgres package.

The test proxy is updated to support the @latest endpoint.

Updates b/126597197

Change-Id: Ic62df9e28b797dafcf73f291adad9282da7911b4
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/535814
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>