internal/stdlib: package for handling the Go standard library
Add a new package containing all code for treating the standard
library as a special module, "std".

Includes a constant for the module path, the function to convert
semvers to Go tags, and a new function for building a zip file from
the Go repo.

Note: adding the go-git package also updated some
other packages in go.mod.

Updates b/138649628.

Change-Id: I0fc2fe43a37825e722b8e635a39f4db3d7350766
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/538930
Reviewed-by: Julie Qiu <julieqiu@google.com>