internal/frontend: replace module tab with subdirectories tab
The module tab on the packages view is replaced with a subdirectories
tab, which only contains packages in that package path directory.

fetchModuleDetails is replaced with fetchPackageDirectoryDetails and
fetchModuleDirectoryDetails, which handle the subdirectories tab in the
packages view and packages tab in the modules view respectively.

The title for the module page is now constructed using the moduleTitle
func. All instances of "std" are replaced with "Standard library".

Fixes b/139371789
Updates b/140485028

Change-Id: Ie0ce25263c0ec6f3979445a51487eef37e5c8914
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/543274
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>