internal/frontend: fix directory view
A bug is fixed that resulted from processPackageOrModulePath, which
prevented the directory view from rendering.

Previously, processPackageOrModulePath return the empty strings if the
code was not http.StatusOK, which resulted in serveDirectoryPage being
executed with path and version set to the empty string. This is fixed by
splitting out parsePathAndVersion from processPackageOrModulePath.

processPackageOrModulePath is also renamed to fetchPackageOrModule,
since it's primary functionality is now to fetch one of those from the
datasource.

Change-Id: I61937c20b37eefb62b75ac292fdee9cd9ae425e5
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/547169
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>