internal/index: make GetVersions fail fast
Previously, a JSON decoding error was logged but didn't stop
processing: GetVersions still returned whatever versions it had
managed to decode from the index.

In the unlikely event that some entry in the middle was corrupted,
we would skip it and never return to it, since we fetch from the
index since the timestamp of the last entry we saw.

It seems safer to fail completely if there are any problems decoding
the index's output.

Also, remove unnecessary UTC conversion, and remove obsolete comment.

Change-Id: Ia29f010a9f31118506c9391a449e68a76e8837f0
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/551716
Reviewed-by: Robert Findley <rfindley@google.com>