internal/derrors: rename NotAcceptable to BadModule
No reason to use an existing HTTP status for this.

Also, pick an unassigned integer in the 4xx range, to avoid any confusion
with an HTTP status.

Also, redo the the error-to-status conversion functions with a table,
to make sure they stay aligned.

Change-Id: Ib9848c2a257f670cac8fd703b99a5c55588cbdc0
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/555197
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>