internal/etl: ignore package if .go file has NULs
build.Context.MatchFile returns an error if the .go file has
bad contents, like zero bytes.

Previously, we did not wrap this error in BadPackageError, so the fetch
failed with a 500 and was retried.

Now we do wrap it, so we skip the package with the bad file and keep
processing.

Fixes b/141427861.

Updates b/133187024.

Change-Id: I9b0d7b96b1462014263bc0757f37a6c97b96880f
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/556719
Reviewed-by: Julie Qiu <julieqiu@google.com>