internal/frontend: remove redundant "Versions"
Don't display the word "Versions" on the versions tab when all
versions belong to a single module.  It's redundant with the tab
title.

Leave everything as is for the multi-module case. This still leaves
nested h2 tags, but that doesn't break anything, and it's rare.

Fixes b/139371418.

Change-Id: Ie69c2337323446069e989688aba9ba8e0b1a1990
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/556726
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>