internal/postgres: add a fast(er) search
A new method is added to postgres.DB, optimistically named 'FastSearch',
that implements search hedging. Each FastSearch concurrently queries
both all packages as well as popular subsets of packages, so that
results can be returned quicker when they consist entirely of popular
packages.

Additionallly SearchResult.Rank is renamed to 'Score', to differentiate
it from ts_rank.

The old 'Search' method is kept for now, although 'log' is corrected to
'ln' (this should have little effect on the ordering of results). In a
subsequent CL we will switch the default search method to FastSearch,
and implement the 'of many' qualifier in the search result summary.

Updates b/141182438

Change-Id: I73c6700e09288f3a8bfdb5fa57d0899606e452d2
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/558516
Reviewed-by: Jonathan Amsterdam <jba@google.com>