migrations: add support functions for hyperloglog estimation
The postgres function 'hashtext' is internal, and therefore really
shouldn't be relied upon. Also, it is only a 32 bit hash.

To work around this, we add a UDF implementing a 64 bit numeric hash
(taking the first 64 bits of the built-in md5 hash), and another that
can be used to count the number of leading zeros in a bigint.

Updates b/141182438

Change-Id: I33051273a41843370fec72b76bed37482617a5b3
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/561155
Reviewed-by: Jonathan Amsterdam <jba@google.com>