internal/source: a package for references to source
Add internal/source, which will know how to generate URLs that
refer to the source code or other files inside modules.

This package will be used to link to module and package source code,
and also for relative links in README files.

For better testing, introduce the httpreplay package. With -record, it
makes HTTP requests to the actual servers and records the responses.
With no flags, it replays the recordings.

Fixes b/141771930.

Updates b/140431331.

Change-Id: I72dcd5535b8be7c6b27dd89245a048d23fd14160
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/561217
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>