internal/source: match go-source templates against known patterns
It's possible that a vanity import path's meta tags mention
a known repo pattern only in the URL templates. This is what gopkg.in
tags do. For example:

  <meta name="go-import" content="gopkg.in/yaml.v2 git https://gopkg.in/yaml.v2">
  <meta name="go-source" content="gopkg.in/yaml.v2 _
        https://github.com/go-yaml/yaml/tree/v2.2.3{/dir}
        https://github.com/go-yaml/yaml/blob/v2.2.3{/dir}/{file}#L{line}">

In this case, we can match the templates in the tag to our list of
known patterns, and use the version-aware templates we get from that
match instead of the ones in the tag. Here, for example, we can note
that the source lives on GitHub and use the pattern that has
"{commit}" instead of the hardcoded "v2.2.3", making it possible to
refer to other versions of the module.

Change-Id: Ia998f47ec52996eb7f71377624c638bedfe1994f
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/562704
Reviewed-by: Julie Qiu <julieqiu@google.com>