internal/postgres: add a searcher that calls popular_search
The previous popularSearcher(cutoff) search methods relied on partial
indexes to pre-filter search results to popular documents, and then
used the boundaries of our scoring function to determine whether such
partial searches were complete. This was a lot faster for popular search
terms, but still left something to be desired:
 + The partial index thresholds (8 importers and 50 importers) required
   quite a lot of analysis to calibrate.
 + Postgres was not able to use the search document GIN index in the
   search, so these popular searches were essentially scanning all
   documents in their popular index. This resulted in a lot of extra
   scanning when search results could have been satisfied by an even
   smaller search.

In order to address these problems, we move away from declarative
queries (and playing games with the query planner), by using the
popular_search stored function to scan packages in descending order of
popularity, and exit as soon as possible.

Along the way, a few other trivial search methods are added to
optionally compare against the (now default) FastSearch.

An off-by-one error is fixed in hyperloglog estimation: we were
operating on the number of leading zeros, rather than the index of first
non-zero bit.

Updates b/141182438

Change-Id: Ib4c6bf16a456e603082d9c9e5c6d430e88879dbb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/564976
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>