internal/etl: mark modules with unprocessed packages
If a module has a package that has .go files, but we can't process the
package because of limitations of our system, mark that fact in the
module_version_states table with a special status.

We choose a 2xx status because the module processing succeeded, but
not 200 so we can distinguish these modules later for reprocessing
if our limitations change (e.g. if we add a build environment).

We choose a non-existent HTTP status (290) to avoid confusion
with any existing status code.

Updates b/133187024.

Change-Id: I1ec1ce7a6d59287c41fcd4d7dacf9ca4b8fd5f2c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/564216
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>