internal/frontend: add forward-progress check in breadcrumbPath
This change ensures there is no situation where the loop never exits.

Previously, this happened when the following call with invalid
parameters was made:

	breadcrumbPath("golang.org/x/tools/notexist", "", "v0.0.0-20191002234911-9ade4c73f2af")

modPath is not expected to be the empty string. It happened due to
another problem elsewhere in the code (which was fixed in CL 564572).

Also improve the verb tense in a log message.

Change-Id: Id93f68da90c5285333b9f642de124bfe090e181c
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/564571
Reviewed-by: Jonathan Amsterdam <jba@google.com>