internal/{derrors,proxy,proxydatasource}: remove Gone category, treat it like Not Found
Module proxies may use 404 and 410 status codes interchangeably
to mean that the proxy does not have the module.¹

Remove the derrors.Gone error category, because at this time there
are no situations in which we want to treat it differently than
Not Found. Should that change in the future, we can re-add it.

Make the mapping from HTTP response status codes to errors more
explicit in the proxy client, since its needs are more specific
than the general derrors.FromHTTPStatus function provides.

¹ https://golang.org/cmd/go/#hdr-Module_downloading_and_verification

Change-Id: I47d4a736f81cebc345d5d0b8125d766d371a0349
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/565300
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>