content,internal/frontend: update header design and add latest badge
The package and module headers are updated based on updated beta designs
at: go/go-discovery-gallery

A badge is added to the package header to indicate if it is the latest
version.

This adds an additional query when the version is not known to be the
latest version. Performance optimizations for handling this will be made
in a future CL.

Fixes b/139411794

Change-Id: Id1ff4e2a3688bf29e0c0447240dc03f3034fce7b