internal/etl: record excluded modules in module_version_states
If a module is on the excluded list when we fetch it, make sure we
record that fact in the module_version_states table. That way the
table has a row for every module we fetch from the index.

Previously we immediately exited from fetching without writing
anything to module_version_states, so if we un-excluded the module
later, we wouldn't have known to reprocess it.

Change-Id: Ic4ad38df8a9f5d0bf776ead6caca87e4a5d38806
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/570139
Reviewed-by: Robert Findley <rfindley@google.com>