migrations: add tsv_parent_directories to packages and search_documents
The following columns are added, which will be used to improve
performance for the directory view:

* search_documents.tsv_parent_directories
* packages.tsv_parent_directories

The function to_tsvector_parent_directories is added, which computes all
directories between a given package path and module path, includes of
the module and package paths. It returns the result as a tsvector.

The following triggers are added:
* trigger_modify_packages_tsv_parent_directories
* trigger_modify_search_documents_tsv_parent_directories

These triggers use to_tsvector_parent_directories to populate the
tsv_parent_directories columns whenever a row is inserted into the
packages or search_documents tables.

Updates b/142392929

Change-Id: If603198ba8a183949b05628d7fd2bf69c0963092
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/569016
Reviewed-by: Robert Findley <rfindley@google.com>