internal/middleware: log request entries at INFO level
Set the severity of the request log entries to Info so that they
appear in the log viewer when the level selector is Info.

Previously, they were at default severity and did not appear, meaning
that if you didn't want to see the debug logs, you also couldn't see
the request entries.

Change-Id: Id4f9b9596c88877eb3332392a3b3df1307ba8791
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/570665
Reviewed-by: Robert Findley <rfindley@google.com>
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>