internal/postgres: make smaller bulk insert statements
Change the maximum number of parameters in a bulk insert statement
to 1000.

We were allowing up to 64K parameters in a bulk insert statement.
Sometimes that was causing the App Engine instance to crash (for
example, the Azure SDK for Go, which has over 1600 packages).

Also, use execTx to improve log output.

Change-Id: Id7c502c9e855cdd9b340ce96a682d486e13ee7f3
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/570666
Reviewed-by: Robert Findley <rfindley@google.com>