internal/postgres: sort to avoid deadlock
Two transactions can deadlock if they insert the same rows into a
table in a different order. That's because when doing an insert,
Postgres acquires a non-exclusive table lock but exclusive row locks.

Before this CL, that actually happened a fair amount with
imports_unique. We haven't seen it with other tables, but it could
happen.

So we sort everything for which there are multiple inserts: packages,
licenses and imports. That way, identical rows are always inserted in
the same order, and deadlock can't happen.

I'm pretty sure this works because I just processed a backlog of about
40,000 modules in dev with no deadlocks.

Fixes b/141164828.

Change-Id: Id8a537fa92efa2a0fbdb5a3cb2e05a46d2afd4f0
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/571784
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>