migrations: add column for storing  source information
Add a column to versions for storing information about where to find
the source code of a module.

We store this information as a single column in JSON format, so we can
easily change the representation without having to migrate the
database.  For example, if we decide to add a new kind of URL template
to distinguish tags from commit hashes, we can add a field to the JSON
without touching the database.

We use the jsonb type because it's more efficient to represent and can
be indexed.  But we don't index it, at least for now.

Change-Id: I4a16740aa9af2ccfba33e5efc3b70a003f8e5d19
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/573456
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>