all.bash: don't use -count=1 by default
Tests are still fast enough not to warrant engineering effort on
speeding them up... but it can be annoying to have a 20-30s delay before
entering a commit message. I find myself more and more frequently
running `git commit --amend -n` to skip the hook.

So let's just allow for cached test results. It might miss some
flakiness, but that should be captured by the CI.

Change-Id: I4cd2c25849997499237fabfc9328ae5a7420911d
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/575058
Reviewed-by: Julie Qiu <julieqiu@google.com>