internal/etl/dochtml: omit Overview section when it's empty
The Overview section displays the package comment and package examples,
if any. All other sections are left out when they have no content. This
change applies the same behavior to the Overview section, displaying it
only when there is content.

Add a test for this by creating a package with no package comment and
no package examples.

Updates b/140266266

Change-Id: I4205b439699c5c56fa91e95b68b5cfbe50bf8a71
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/575435
Reviewed-by: Julie Qiu <julieqiu@google.com>