internal/middleware: add a page-caching middleware
A page caching middleware is added that caches any 200 OK responses for
a configurable duration. It is instrumented with custom metrics.

For now, this isn't wired into the frontend, though the code to do so is
commented out. Tests will be added in a follow-up CL.

Updates 143163521

Change-Id: Id919d3e63e1b43f5905e7f20df475a3c69379e87
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/580776
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>