internal/frontend: let the user copy the import path
Add a "copy" icon next to our breadcrumb path that lets
the user copy the package or module path to the clipboard.
This supports:

- Pasting a package path into a Go import declaration;
- Pasting a module path into a go.mod file.

The appearance is based on the mock


Complicated by the case where the module is the standard library itself;
then there's nothing to copy, so we don't show the icon.

Change-Id: I3827424a105ea4fa9e941d3e56b5424451b19f23
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/579359
Reviewed-by: Alexander Rakoczy <rakoczy@google.com>
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>