migrations: set to_tsv_parent_directories, hll_hash and hll_zeros as parallel safe
These functions are set as PARALLEL SAFE, since they do not modify ay
database state and are safe to be run in parallel:
* hll_hash
* hll_zeros
* to_tsvector_parent_directories

The UPDATE statements in migration 000002 are removed since they failed
to finish after several hours in dev. Although we expect this to run
much faster with PARALLEL SAFE set, moving these statements to a script
just in case.

Fixes b/143294693

Change-Id: I396073af2f408143287ce9c85b9dcd78e46ef213
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/583335
Reviewed-by: Jonathan Amsterdam <jba@google.com>