internal/postgres: set the postgres statement_timeout parameter
Set Postgres's statement_timeout parameter.  This should prevent
runaway statements.  We have to set it each time we open a DB
connection.

Also, the timeout on the fetch context should be related to the
statement timeout, so set it to twice the value.

Change-Id: I9568ae81b8d2bf86c7f642d5de530acfa7fa90e7
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/583658
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>