internal/frontend: improve copy button look and feel
Make the copy-import-path button into a proper button. This makes it
part of the tab order, correctly highlights it when it has focus, and
makes the special keypress event handlers unnecessary.

The only way I could get the icon to display in medium turquoise was
to set the fill directly on the svg element (and remove overriding
fill attributes on sub-elements).

Change-Id: I3c16a4f99a6b7d41ad8df096de02680090c1a16d
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/584335
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Alexander Rakoczy <rakoczy@google.com>