internal/etl/dochtml: move html for htmlPackage to a separate .go file
The HTML template for htmlPackage is moved to a separate .go file to
make it easier to edit the template. Ideally, this would be move to a
.tmpl file, but this would require parsing the template in etl.NewServer
and passing *template.Template all the way down to loadPackage, which is
a bigger change than we want to make at this time.

Additionally, we expect the internal/etl/dochtml to change soon as
progress is made on godoc libraries.

Updates b/137567588

Change-Id: I05ccf15b1020fb8bcc9c232349bddea0f8e59eeb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/588352
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>