frontend,middleware: longer caching for details pages
Now that the latest badge is asynchronous, details content is
essentially static and can be cached for significantly longer.
Unfortunately, this is not true for all tabs, so the cache TTL
calculation must be made dynamic.

This CL modifies the caching middleware to support TTLs that depend on
the incoming request, and uses this functionality to implement
significantly longer caching for details tabs other than 'versions' and
'importedby'.

Change-Id: I469c3727bf08cd7b7d1dec1b171ddd2f6307ad6f
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/589584
Reviewed-by: Jonathan Amsterdam <jba@google.com>