internal/integration: add an initial integration test
We're about to need better testing around the interaction between the
frontend, etl, and redis cache. For example: to be able to validate
things like the 'latest' badge being in sync with the latest page, or to
verify that the cache is invalidated when we fetch a new version.

This CL is a starting point for such testing, adding a package for
integration testing and a simple end-to-end test involving processing
the index and serving a module's documentation.

Some inconsistency in the index test client was corrected.

Updates b/143761646
Updates b/143760329

Change-Id: I648dbe4e49656d7132374d773f5a590a15f775a9
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/589967
Reviewed-by: Julie Qiu <julieqiu@google.com>