internal/frontend: extend /latest-version for packages
Previously, the /latest-version endpoint returned the latest module
version. This wasn't always correct for packages: the latest version
of a package might not be the latest version of its module (since it
might have been removed from the module in later versions).

Example: github.com/docker/docker is both a package and a module.
- The package's latest version is v0.9.1.
- The module's latest version is v1.13.1.

Visiting /github.com/docker/docker takes you to v0.9.1. As of this CL,
the badge says "Latest".  Previously it said "Go to latest," but the
link took you back to the same page.

Fixes b/143833787.

Change-Id: I9600ac7e1db22ae0c4168e310a633f757dae9cd6
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/590622
Reviewed-by: Robert Findley <rfindley@google.com>