internal/frontend: functions for version strings
We transform versions in two ways: for display, and for linking.
In both cases the stdlib is special.

Provide functions for each, and use them throughout.

Also, change a panic to a log message, just in case.

Fixes b/143863980.

Tested:
- On master:
  > devtools/run_local.sh dev frontend &
  > devtools/capture_pages.sh /tmp/master http://localhost:8080 'context?tab=versions' 'mod/std?tab=versions' 'github.com/pkg/errors?tab=versions' context@go1.13.5

- On this branch:
  > devtools/run_local.sh dev frontend &
  > devtools/capture_pages.sh /tmp/mybranch http://localhost:8080 'context?tab=versions' 'mod/std?tab=versions' 'github.com/pkg/errors?tab=versions' context@go1.13.5

> diff /tmp/master /tmp/mybranch

The only difference was the error message in the last page.

Change-Id: Iba7089ff2ae3af96f9e2b1b67bc18cae8c8c6d70
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/592079
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>