internal/frontend: change URLS on details page when latest version is requested
The URL logic is updated so that when a user requests a
package/module/directory with /<path> and do not specify a version, the
links to the tabs also follow the /path?tab=<tab> scheme.

The only exception to this is packages on the directories tab, because
(1) it required more extensive refactoring and (2) the latest version of
a package might not be part of the same module, so deep linking seemed
preferable.

Change-Id: I7bc92076ccb59b933c752377dd764128290228c4
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/592720
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>