content: include version in latest link
Because of cache timings, it's possible for unversioned details pages to
point to a different version than the /latest-version endpoint. This
could result in the poor experience where the user clicks on the 'Go to
latest' badge, and is taken to a page that still contains a 'Go to
latest' badge.

This CL changes the 'Go to latest' badge to explicitly link to the
version returned by the /latest-version endpoint, in order to avoid the
situation described above. It's still possible for the caches to go out
of sync, but I think this makes it less visible.

Fixes b/143761646

Change-Id: If07d94fe5939fc59e076aa58d26a3c5c1c098ffb
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/593056
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>