content,internal/frontend: redirect URLs without tab param to doc
When a user visits a URL that does not have the tab query param
specified, they will now be redirected to the doc tab. Previously, the
user was shown the doc tab without being redirected.

Additionally, always show the repository URL and module page link on the
overview tab, regardless of the license. When a package/module is not
redistributable, the user is now shown the overview tab.

Updates b/143953613

Change-Id: I655cca597c6f31accf4f8f2806a442a6588cb692
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/594571
Reviewed-by: Jonathan Amsterdam <jba@google.com>