internal/frontend: TestServer matches html structure
Previously, the TestServer function fetched pages and treated the HTML
body as a string, looking for literal substrings. With this CL, it
parses the HTML and looks for patterns in the structure.

This makes the tests more precise and more flexible. We can still
get the old behavior by writing

    in("", text("something"))

(which means: in the entire body, look for the string "something").

But we can also, for instance, test which tab is selected, by writing

    in("li.selected", text("Imports"))

The previous tests were brittle in that small, irrelevant changes could break them.
For example, a test for the literal string

    <a href="foo">bar</a>

would break if, say, a class were added to the tag:

   <a class="c" href="foo">bar</a>

Now we can ignore other attributes by writing

    in("a", href("foo"), text("bar"))

One thing we lose is the ability to say that one string must occur
after another; the old list of strings had to occur in the order they
appeared. But that is more than made up for by the ability to focus on
a particular element or subtree when looking for a string.

This CL also fixes a bug in the test where the module URL was not
being checked.

PRODUCTION RISK: none
Only tests are changed.

Change-Id: I8304c279df25f32a4a2b669a6bb6a17fa7baef44
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/597500
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>